Move pointer increments to their own line
It was mentioned that some compilers - such as the PS4 compiler - will
complain when the increment is in some other line. This change doesn't
really help outside of removing those potential problems. It's unlikely
that I will ever compile this for the PS4, but it's a habit I want to
form.